Remove file size limits

+++ src/object.rs
@@ -11,12 +11,12 @@ use std::{
 
 #[derive(Debug, Clone, PartialEq, Eq, Copy)]
 pub struct Span {
-    pub start: (u16, u16),
-    pub end: (u16, u16),
+    pub start: (usize, usize),
+    pub end: (usize, usize),
 }
 
 impl Span {
-    pub fn new(start: (u16, u16), end: (u16, u16)) -> Self {
+    pub fn new(start: (usize, usize), end: (usize, usize)) -> Self {
         Span { start, end }
     }
 }
